refactor(squarem): move parameter packing into the stats layer; regenerate the API manifest
Two of this repo's own guards caught the first cut, both fixed as designed rather than suppressed:

- compute_metadata_test bars inference/em.py from naming concrete distribution families (its module
  docstring promises orchestration only). squarem_packer + the family handlers now live in
  mixle/stats/parameter_packing.py -- the stats layer, where concrete families are legal -- and
  em.py imports only the packer FUNCTION. The new module's own top level imports nothing but numpy
  (handlers import lazily), so it joins the guard's whitelist with that rationale written in place.
- public_api_manifest_test flagged the two intended new names (SquaremEM, squarem_packer) --
  api_manifest.json regenerated to declare them.
