+++ api_manifest.json
@@ -591,6 +591,7 @@
     "Simulator",
     "Skill",
     "SkillRegistry",
+    "SquaremEM",
     "StreamingEstimator",
     "TestResult",
     "TrialsToTarget",
@@ -776,6 +777,7 @@
     "spatial_block_kfold",
     "split_conformal",
     "split_rhat",
+    "squarem_packer",
     "stouffer_combine",
     "stratified_kfold",
     "synthesize",
